WebExtensions with package.json should not be classified as JetPack addon
This fixes a bug with WebExtensions that have both a `package.json`
and a `manifest.json`. This is very common and `firefox-profile`
classifies these extensions as a JetPack style addon. Firefox subsequently
fails to load them.

JetPack style addons are deprecated and we should first attempt to
read the manifest.json in case this is a WebExtension.

An example addon for which the old approach is problematic is AdBlock Plus.
It lists both `package.json` and `manifest.json`.